+++ openrouteservice/src/main/java/org/heigit/ors/routing/algorithms/AbstractManyToManyRoutingAlgorithm.java
@@ -17,11 +17,13 @@ import com.graphhopper.routing.util.DefaultEdgeFilter;
 import com.graphhopper.routing.util.EdgeFilter;
 import com.graphhopper.routing.util.FlagEncoder;
 import com.graphhopper.routing.util.TraversalMode;
+import com.graphhopper.routing.weighting.TurnWeighting;
 import com.graphhopper.routing.weighting.Weighting;
 import com.graphhopper.storage.Graph;
 import com.graphhopper.storage.NodeAccess;
 import com.graphhopper.util.EdgeExplorer;
 import com.graphhopper.util.EdgeIterator;
+import org.heigit.ors.config.MatrixServiceSettings;
 
 public abstract class AbstractManyToManyRoutingAlgorithm implements ManyToManyRoutingAlgorithm {
 	protected final Graph graph;
@@ -63,7 +65,11 @@ public abstract class AbstractManyToManyRoutingAlgorithm implements ManyToManyRo
 	}
 
 	protected boolean accept(EdgeIterator iter, int prevOrNextEdgeId) {
-		return additionalEdgeFilter == null || additionalEdgeFilter.accept(iter);
+		if (MatrixServiceSettings.getUTurnCost() == TurnWeighting.INFINITE_U_TURN_COSTS && iter.getEdge() == prevOrNextEdgeId) {
+			return false;
+		} else {
+			return additionalEdgeFilter == null || additionalEdgeFilter.accept(iter);
+		}
 	}
 
 	@Override