+++ openrouteservice/src/main/java/org/heigit/ors/routing/graphhopper/extensions/storages/builders/WaySurfaceTypeGraphStorageBuilder.java
@@ -62,7 +62,7 @@ public class WaySurfaceTypeGraphStorageBuilder extends AbstractGraphStorageBuild
 		int surfaceType = way.hasTag(TAG_SURFACE) ? SurfaceType.getFromString(way.getTag(TAG_SURFACE)) : SurfaceType.UNKNOWN;
 		if (surfaceType == SurfaceType.UNKNOWN) {
 			if (wayType == WayType.ROAD || wayType == WayType.STATE_ROAD || wayType == WayType.STREET) {
-				surfaceType = SurfaceType.ASPHALT;
+				surfaceType = SurfaceType.PAVED;
 			} else if (wayType == WayType.PATH) {
 				surfaceType = SurfaceType.UNPAVED;
 			}